Add tests for further coverage (model). Move CardModel tests into their own file.
